rag_requirements.txt
gradio
pinecone
python-dotenv
plotly
pandas
scikit-learn
google-genai
